AB#- 10710 - Changes to implement group deletion and refactored grouping code by creating service
