+++ Tests/UnitTests/fss-search.component.spec.ts
@@ -14,6 +14,7 @@ import { AppConfigService } from '../../src/app/core/services/app-config.service
 import { FssSearchService } from '../../src/app/core/services/fss-search.service';
 import { FssSearchFilterService } from '../../src/app/core/services/fss-search-filter.service';
 import { Field, FssSearchRow, Operator } from '../../src/app/core/models/fss-search-types';
+import { FssSearchGroupingService } from '../../src/app/core/services/fss-search-grouping.service';
 
 describe('FssSearchComponent', () => {
   let component: FssSearchComponent;
@@ -22,6 +23,7 @@ describe('FssSearchComponent', () => {
   let searchFilterservice: FssSearchFilterService;
   let searchService: FssSearchService;
   let elementRef: ElementRef;
+  let searchGroupingService: FssSearchGroupingService;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
@@ -44,6 +46,7 @@ describe('FssSearchComponent', () => {
     searchService = TestBed.inject(FssSearchService);
     searchFilterservice = TestBed.inject(FssSearchFilterService);
     fileShareApiService = TestBed.inject(FileShareApiService);
+    searchGroupingService = TestBed.inject(FssSearchGroupingService);
   });
 
   it('should create', () => {
@@ -53,7 +56,7 @@ describe('FssSearchComponent', () => {
   });
 
   test('should return datatype as "attribute" when UserAttribute is passed', () => {
-    component = new FssSearchComponent(searchService, searchFilterservice, fileShareApiService, elementRef);
+    component = new FssSearchComponent(searchService, searchFilterservice, fileShareApiService, elementRef, searchGroupingService);
     component.ngOnInit();
     component.fields = searchService.getFields(MockUserAttributeFields());
     console.log(component);
@@ -61,7 +64,7 @@ describe('FssSearchComponent', () => {
   });
 
   test('should return datatype as "string" when SystemAttribute FileName is passed', () => {
-    component = new FssSearchComponent(searchService, searchFilterservice, fileShareApiService, elementRef);
+    component = new FssSearchComponent(searchService, searchFilterservice, fileShareApiService, elementRef, searchGroupingService);
     component.ngOnInit();
     component.fields = searchService.getFields(MockUserAttributeFields());
     console.log(component);
@@ -69,7 +72,7 @@ describe('FssSearchComponent', () => {
   });
 
   test('should return filtered operators based on number datatype', () => {
-    component = new FssSearchComponent(searchService, searchFilterservice, fileShareApiService, elementRef);
+    component = new FssSearchComponent(searchService, searchFilterservice, fileShareApiService, elementRef, searchGroupingService);
     component.ngOnInit();
     const expectedOperators: Operator[] = [
       { value: 'eq', text: '=', type: 'operator', supportedDataTypes: ['string', 'number', 'date', 'attribute'] },
@@ -84,7 +87,7 @@ describe('FssSearchComponent', () => {
   });
 
   test('should return value type text based on datatype string', () => {
-    component = new FssSearchComponent(searchService, searchFilterservice, fileShareApiService, elementRef);
+    component = new FssSearchComponent(searchService, searchFilterservice, fileShareApiService, elementRef, searchGroupingService);
     component.ngOnInit();
     var expectedValueType = "text";
     var result = component.getValueType('string');
@@ -93,7 +96,7 @@ describe('FssSearchComponent', () => {
   });
 
   test('should return operator type nullOperator based on operator value ne null', () => {
-    component = new FssSearchComponent(searchService, searchFilterservice, fileShareApiService, elementRef);
+    component = new FssSearchComponent(searchService, searchFilterservice, fileShareApiService, elementRef, searchGroupingService);
     component.ngOnInit();
     var expectedOperatorType = "nullOperator";
     var changedOperator = { operatorValue: "ne null", rowId: 1 }
@@ -103,7 +106,7 @@ describe('FssSearchComponent', () => {
   });
 
   test('should return second row based on rowid passed', () => {
-    component = new FssSearchComponent(searchService, searchFilterservice, fileShareApiService, elementRef);
+    component = new FssSearchComponent(searchService, searchFilterservice, fileShareApiService, elementRef, searchGroupingService);
     component.ngOnInit();
     let searchRows: FssSearchRow[] = [];
     let fields: Field[] = [
@@ -157,7 +160,7 @@ describe('FssSearchComponent', () => {
   });
 
   test('should return flag false when FileSize value is passed as number', () => {
-    component = new FssSearchComponent(searchService, searchFilterservice, fileShareApiService, elementRef);
+    component = new FssSearchComponent(searchService, searchFilterservice, fileShareApiService, elementRef, searchGroupingService);
     component.ngOnInit();
     let searchRows: FssSearchRow[] = [];
     let fields: Field[] = [
@@ -180,7 +183,7 @@ describe('FssSearchComponent', () => {
   });
 
   test('should return row count 2 when Add new line is called twice', () => {
-    component = new FssSearchComponent(searchService, searchFilterservice, fileShareApiService, elementRef);
+    component = new FssSearchComponent(searchService, searchFilterservice, fileShareApiService, elementRef, searchGroupingService);
     component.ngOnInit();
     component.fields = searchService.getFields(MockUserAttributeFields());
     component.addSearchRow();
@@ -191,7 +194,7 @@ describe('FssSearchComponent', () => {
   });
 
   test('should return row count 2 when 3 rows exist and delete is called', () => {
-    component = new FssSearchComponent(searchService, searchFilterservice, fileShareApiService, elementRef);
+    component = new FssSearchComponent(searchService, searchFilterservice, fileShareApiService, elementRef, searchGroupingService);
     component.ngOnInit();
     component.fields = searchService.getFields(MockUserAttributeFields());
     component.addSearchRow();
@@ -204,7 +207,7 @@ describe('FssSearchComponent', () => {
   });
 
   test('should disable Value input when nullOperator is selected in the row', () => {
-    component = new FssSearchComponent(searchService, searchFilterservice, fileShareApiService, elementRef);
+    component = new FssSearchComponent(searchService, searchFilterservice, fileShareApiService, elementRef, searchGroupingService);
     component.ngOnInit();
     component.fields = searchService.getFields(MockUserAttributeFields());    
     component.fssSearchRows.push(createSearchRow(1, component.fields, component.operators, 'AND', 'FileSize', 'nullOperator', 'test', 'tel', false));
