+++ src/app/features/fss-search/fss-search.component.ts
@@ -4,6 +4,7 @@ import { FssSearchService } from './../../core/services/fss-search.service';
 import { Operator, IFssSearchService, Field, JoinOperator, FssSearchRow, RowGrouping, GroupingLevel, UIGrouping } from './../../core/models/fss-search-types';
 import { FileShareApiService } from '../../core/services/file-share-api.service';
 import { FssSearchFilterService } from '../../core/services/fss-search-filter.service';
+import { FssSearchGroupingService } from '../../core/services/fss-search-grouping.service';
 
 
 @Component({
@@ -43,7 +44,7 @@ export class FssSearchComponent implements OnInit {
   groupingLevels: GroupingLevel[]=[];
   uiGroupings: UIGrouping[] = [];  
   @ViewChild("ukhoTarget") ukhoDialog: ElementRef;
-  constructor(private fssSearchTypeService: IFssSearchService, private fssSearchFilterService: FssSearchFilterService, private fileShareApiService: FileShareApiService, private elementRef: ElementRef) { }
+  constructor(private fssSearchTypeService: IFssSearchService, private fssSearchFilterService: FssSearchFilterService, private fileShareApiService: FileShareApiService, private elementRef: ElementRef, private fssSearchGroupingService: FssSearchGroupingService) { }
 
   ngOnInit(): void {
     this.joinOperators = this.fssSearchTypeService.getJoinOperators();
@@ -60,7 +61,7 @@ export class FssSearchComponent implements OnInit {
   addSearchRow() {
     this.fssSearchRows.push(this.getDefaultSearchRow());
     this.rowId += 1;
-    this.createUIGrouping();
+    this.uiGroupings = this.fssSearchGroupingService.createUIGrouping(this.fssSearchRows);
   }
 
   getDefaultSearchRow() {
@@ -153,7 +154,11 @@ export class FssSearchComponent implements OnInit {
   }
 
   onSearchRowDeleted(rowId: number) {
-    this.fssSearchRows.splice(this.fssSearchRows.findIndex(fsr => fsr.rowId === rowId), 1);
+    var deleteRowIndex = this.fssSearchRows.findIndex(fsr => fsr.rowId === rowId);    
+    this.fssSearchRows.splice(deleteRowIndex, 1);
+
+    this.rowGroupings = this.fssSearchGroupingService.resetRowGroupings(this.rowGroupings, deleteRowIndex);    
+    this.setupGrouping();
   }
 
   validateSearchInput() {
@@ -322,8 +327,8 @@ onGroupClicked(){
       );
   }
   else{       
-      this.AddGrouping();       
-      this.createUIGrouping(); 
+      this.addGrouping();       
+      this.setupGrouping();
   }
 }
 
@@ -341,166 +346,33 @@ isGroupIntersectWithOther() {
       this.currentGroupEndIndex > g.endIndex)) !== undefined)
 }
 
-AddGrouping(){
-
+addGrouping(){
   this.rowGroupings.push({        
     startIndex: this.currentGroupStartIndex, 
     endIndex: this.currentGroupEndIndex
   });
-  
-  if(this.groupingLevels.length == 0){   
+}
 
-    var groupingLevel = new GroupingLevel();
-    groupingLevel.level = 1;
-    groupingLevel.rowGroupings.push({startIndex: this.currentGroupStartIndex, endIndex: this.currentGroupEndIndex});
-    this.groupingLevels.push(groupingLevel);
-
-  }  
-  else if(this.isOuterLevelGroup()){
-        
-    var matchedGroupingLevel = this.groupingLevels.filter(g=> g.rowGroupings.some( r => 
-        r.startIndex >= this.currentGroupStartIndex && r.endIndex <= this.currentGroupEndIndex));
-
-    var newLevel = new GroupingLevel();
-    var matchedLevel = matchedGroupingLevel[matchedGroupingLevel.length-1];
-    
-    if(matchedLevel !== undefined){
-      var currentlevel = this.groupingLevels.find(g => g.level === matchedLevel.level && g.rowGroupings.find(r=> 
-        r.startIndex <= this.currentGroupStartIndex && r.endIndex >= this.currentGroupEndIndex));      
-      
-      if(currentlevel !== undefined) {
-        newLevel = currentlevel;      
-      }
-      else {
-        newLevel.level = matchedLevel.level + 1;
-      }
-    }
-      
-    var newLevelIndex = this.groupingLevels.findIndex(i => i.level === newLevel.level);           
-    if(newLevelIndex !== -1){
-      this.groupingLevels[newLevelIndex].rowGroupings.push({startIndex:this.currentGroupStartIndex, endIndex:this.currentGroupEndIndex});
-    }
-    else{
-      newLevel.rowGroupings.push({startIndex: this.currentGroupStartIndex, endIndex: this.currentGroupEndIndex});
-      this.groupingLevels.push(newLevel);
-    }
-  }     
-  else if(this.isInnerLevelOfExistingGroup()) {
-        this.showMessage(
-          "info",
-          "Adding an inner group to a group is not supported.",
-          "To add an inner group, first remove the outer group, then add the inner group and re-add the outer group."
+setupGrouping(){
+  var groupingDetails = this.fssSearchGroupingService.resetGroupingDetails(this.rowGroupings,this.fssSearchRows);
+  if(groupingDetails.isInnerGroupError)
+  {
+    this.showMessage(
+            "info", 
+            "Adding an inner group to a group is not supported.",
+            "To add an inner group, first remove the outer group, then add the inner group and re-add the outer group."
         );
-        this.rowGroupings.pop();
-  }    
-  else if(this.isInnerLevelGroup()){  
-
-    var existingGroupingLevel = this.groupingLevels.filter(g => g.rowGroupings.some( r =>    
-      (r.startIndex < this.currentGroupStartIndex && r.endIndex < this.currentGroupEndIndex) ||
-      (r.startIndex > this.currentGroupStartIndex && r.endIndex > this.currentGroupEndIndex )));
-
-      var existingLevel = existingGroupingLevel[0];
-      var existingLevelIndex = this.groupingLevels.findIndex(i => i === existingLevel);        
-      existingLevel.rowGroupings.push({startIndex: this.currentGroupStartIndex, endIndex: this.currentGroupEndIndex});
-
-      this.groupingLevels[existingLevelIndex] = existingLevel;
-  }    
+    this.rowGroupings.pop();
+  }
+  this.groupingLevels = groupingDetails.groupingLevels;
+  this.uiGroupings = groupingDetails.uiGroupings;
 }
 
-isOuterLevelGroup(){
-  var outerGroup =  this.groupingLevels.filter(g=> g.rowGroupings.some( 
-    r => r.startIndex >= this.currentGroupStartIndex && r.endIndex <= this.currentGroupEndIndex));
-
-  if (outerGroup.length>0){
-    return true;
-  }
-  return false;
-}
-
-isInnerLevelOfExistingGroup(){
-  var innerGroup = this.groupingLevels.filter(g => g.rowGroupings.find(
-                  r => r.startIndex <= this.currentGroupStartIndex && r.endIndex >= this.currentGroupEndIndex));
-
-  if (innerGroup.length>0){
-    return true;
-  }
-  return false;
-}
-
-isInnerLevelGroup(){  
-  var innerGroup = this.groupingLevels.find(g => g.rowGroupings.find( r => 
-    ((r.startIndex < this.currentGroupStartIndex && r.endIndex < this.currentGroupEndIndex) && r.startIndex < this.currentGroupEndIndex) ||
-    (r.startIndex > this.currentGroupStartIndex && r.endIndex > this.currentGroupEndIndex )));
-
-    if (innerGroup !== undefined){
-      return true;
-    }
-    return false;
-}
-
-createUIGrouping(){
-  this.uiGroupings = [];
-
-  if(this.groupingLevels.length > 0){
-    var maxLevel = this.groupingLevels[this.groupingLevels.length-1].level;
-
-    for(var i = 0; i < this.fssSearchRows.length; i++){  
-      var j = maxLevel;
-
-      while(j>=1){
-        var groupingLevel = this.groupingLevels.find(g=>g.level === j);               
-        var uiGrouping = new UIGrouping();
-        uiGrouping.rowIndex = i;
-        uiGrouping.class = this.getUIGroupClass(i, groupingLevel!);
-        uiGrouping.colspan = this.getUIGroupingColspan(i, groupingLevel!);
-        uiGrouping.rowGroupings = this.getUIRowGrouping(i, groupingLevel!);
-        this.uiGroupings.push(uiGrouping);  
-        
-        j = j - uiGrouping.colspan; 
-      }          
-    }
-  }
-}
-
-getUIGroupClass(rowIndex: number, groupingLevel: GroupingLevel){
-  var groupingClass = "";
-
-  if(groupingLevel.rowGroupings.find(g=>g.startIndex == rowIndex)){
-    groupingClass = "group group-start";
-  }
-  else if(groupingLevel.rowGroupings.find(g=>g.endIndex == rowIndex)){
-    groupingClass = "group group-end";
-  }
-  else if(groupingLevel.rowGroupings.find(g=>g.startIndex < rowIndex && g.endIndex > rowIndex)){
-    groupingClass = "group";
-  }
-  else{
-    groupingClass = "no-group";
-  }
-
-  return groupingClass;
-}
-
-getUIGroupingColspan(rowIndex:number, groupingLevel: GroupingLevel) {
-  var groupingLevels = this.groupingLevels.slice().reverse()
-       .filter(gl => gl.level <= groupingLevel.level && gl.rowGroupings
-       .some(g => rowIndex >= g.startIndex &&
-        rowIndex <= g.endIndex ));
-
-    if(groupingLevels.length > 1) {
-      return groupingLevels[0].level - groupingLevels[1].level; 
-    } 
-    else if(groupingLevels.length === 1 && groupingLevels[0].level !== groupingLevel.level ) {
-      return groupingLevels[0].level;
-    }
-    else {
-      return groupingLevel.level;
-    }      
-}
-
-getUIRowGrouping(rowIndex: number, groupingLevel:GroupingLevel){
-  var rowGrouping = groupingLevel.rowGroupings.filter(r=>r.startIndex == rowIndex);
-  return rowGrouping;
+onGroupDeleted(rowGrouping: any) { 
+  var start = rowGrouping.rowGrouping[0].startIndex;
+  var end = rowGrouping.rowGrouping[0].endIndex;
+  this.rowGroupings.splice(this.rowGroupings.findIndex(r => r.startIndex === start && r.endIndex === end),1);  
+  this.setupGrouping();  
 }
 
 }
